Fix - EncryptionKey não sendo adicionada ao entrar na partida | Add - Traduções FTL
