+++ Content.Server/Administration/Commands/SetOutfitCommand.cs
@@ -12,9 +12,7 @@ using Content.Shared.Roles;
 using Robust.Shared.Console;
 using Robust.Shared.Player;
 using Robust.Shared.Prototypes;
-using Content.Shared.Radio.Components; // Parkstation-IPC
-using Content.Shared.Containers; // Parkstation-IPC
-using Robust.Shared.Containers; // Parkstation-IPC
+using Content.Server.Silicons.IPC;
 
 namespace Content.Server.Administration.Commands
 {
@@ -129,36 +127,7 @@ namespace Content.Server.Administration.Commands
                     handsSystem.TryPickup(target, inhandEntity, checkActionBlocker: false, handsComp: handsComponent);
                 }
             }
-
-            // Parkstation-IPC-Start
-            // Pretty much copied from StationSpawningSystem.SpawnStartingGear
-            if (entityManager.TryGetComponent<EncryptionKeyHolderComponent>(target, out var keyHolderComp))
-            {
-                var earEquipString = startingGear.GetGear("ears", profile);
-                var containerMan = entityManager.System<SharedContainerSystem>();
-
-                if (!string.IsNullOrEmpty(earEquipString))
-                {
-                    var earEntity = entityManager.SpawnEntity(earEquipString, entityManager.GetComponent<TransformComponent>(target).Coordinates);
-
-                    if (entityManager.TryGetComponent<EncryptionKeyHolderComponent>(earEntity, out _) && // I had initially wanted this to spawn the headset, and simply move all the keys over, but the headset didn't seem to have any keys in it when spawned...
-                        entityManager.TryGetComponent<ContainerFillComponent>(earEntity, out var fillComp) &&
-                        fillComp.Containers.TryGetValue(EncryptionKeyHolderComponent.KeyContainerName, out var defaultKeys))
-                    {
-                        containerMan.CleanContainer(keyHolderComp.KeyContainer);
-
-                        foreach (var key in defaultKeys)
-                        {
-                            var keyEntity = entityManager.SpawnEntity(key, entityManager.GetComponent<TransformComponent>(target).Coordinates);
-                            keyHolderComp.KeyContainer.Insert(keyEntity, force: true);
-                        }
-                    }
-
-                    entityManager.QueueDeleteEntity(earEntity);
-                }
-            }
-            // Parkstation-IPC-End
-
+            InternalEncryptionKeySpawner.TryInsertEncryptionKey(target, startingGear, entityManager, profile); // Parkstation - IPC
             return true;
         }
     }
