+++ Resources/Locale/pt-BR/power/batteries.ftl
@@ -1,0 +1,1 @@
+battery-electrocute-charge = A bateria tomou um disparo de energia!
