+++ Resources/Locale/pt-BR/power/batteryDrinker.ftl
@@ -1,0 +1,2 @@
+battery-drinker-verb-drink = Drenar
+battery-drinker-empty = {CAPATALIZE(THE($target))} já está descarregado!
