+++ Resources/Locale/pt-BR/silicons/silicons.ftl
@@ -1,0 +1,5 @@
+silicon-overheating = Você se sente seus circuitos torrarem!
+silicon-crit = Integridade estrutural em estado critico!
+silicon-power-low = Pouca bateria!
+alerts-charge-name = Carregar
+alerts-charge-desc = Seu nível atual de bateria.
