+++ Resources/Locale/pt-BR/silicons/siliconsChargers.ftl
@@ -1,0 +1,4 @@
+silicon-charger-overheatwarning = Você se sente seus circuitos torrarem!
+silicon-charger-chargerate-string = Nível de carga
+silicon-charger-efficiency-string = Eficiência
+silicon-charger-list-full = {CAPITALIZE(THE($charger))} só pode lidar com alguns alvos!
