+++ Resources/Locale/pt-BR/species/IPC.ftl
@@ -1,0 +1,1 @@
+species-name-ipc = IPC
