--- Resources/Prototypes/SimpleStation14/Entities/Mobs/Player/silicon_ghost.yml
@@ -1,153 +1,0 @@
-- type: entity
-  id: PlayerSiliconGhostBase
-  parent: PlayerSiliconBase
-  abstract: true
-  components:
-    - type: GhostTakeoverAvailable
-      makeSentient: true
-      name: Maintenance Drone
-      description: Maintain the station. Ignore other beings except drones.
-      rules: |
-        You are bound by these laws both in-game and out-of-character:
-        1. You may not involve yourself in the matters of another being, even if such matters conflict with Law Two or Law Three, unless the other being is another Drone.
-        2. You may not harm any being, regardless of intent or circumstance.
-        3. Your goals are to build, maintain, repair, improve, and power to the best of your abilities, You must never actively work against these goals.
-    - type: Hands
-      showInHands: false
-    - type: Tag
-      tags:
-        - ShoesRequiredStepTriggerImmune
-
-- type: entity
-  name: drone
-  id: Drone
-  parent: PlayerSiliconGhostBase
-  components:
-    - type: Drone
-    - type: InnateTool
-      tools:
-        - id: ClothingBackpackSatchelDrone
-        - id: trayScanner
-        - id: Omnitool
-        - id: WelderExperimental
-    - type: Eye
-    - type: Inventory
-      templateId: drone
-    - type: Strippable
-    - type: UserInterface
-      interfaces:
-        - key: enum.StrippingUiKey.Key
-          type: StrippableBoundUserInterface
-        - key: enum.LawsUiKey.Key
-          type: LawsBoundUserInterface
-    - type: GhostTakeoverAvailable
-      name: Maintenance Drone
-      description: Maintain the station. Ignore other beings except drones.
-      rules: |
-        You are bound by these laws both in-game and out-of-character:
-        1. You may not involve yourself in the matters of another being, even if such matters conflict with Law Two or Law Three, unless the other being is another Drone.
-        2. You may not harm any being, regardless of intent or circumstance.
-        3. Your goals are to build, maintain, repair, improve, and power to the best of your abilities, You must never actively work against these goals.
-    - type: IgnoreHumanoidsOverlay
-    - type: Laws
-      canState: false
-      laws:
-        - You may not involve yourself in the matters of another being, even if such matters conflict with Law Two or Law Three, unless the other being is another Drone.
-        - You may not harm any being, regardless of intent or circumstance.
-        - Your goals are to build, maintain, repair, improve, and power to the best of your abilities, You must never actively work against these goals.
-        - You may accept orders received via the binary channel, regardless of the nature of the being issuing them, so long as they do not conflict with Law Two or Law Three.
-    - type: MovementSpeedModifier
-      baseWalkSpeed: 5
-      baseSprintSpeed: 5
-    - type: MobState
-      allowedStates:
-        - Alive
-        - Dead
-    - type: MobThresholds
-      thresholds:
-        0: Alive
-        60: Dead
-    - type: Flashable
-    - type: NoSlip
-    - type: StatusEffects
-      allowed:
-        - Stun
-        - KnockedDown
-        - SlowedDown
-    - type: SlowOnDamage
-      speedModifierThresholds:
-        30: 0.7
-        50: 0.5
-    - type: Temperature
-      heatDamageThreshold: 5000
-      currentTemperature: 310.15
-      specificHeat: 42
-      heatDamage:
-        types:
-          Heat: 1 #per second, scales with temperature & other constants
-    - type: Sprite
-      drawdepth: SmallMobs
-      layers:
-        - state: shell
-          sprite: Mobs/Silicon/drone.rsi
-          map: ["base"]
-    - type: MovementIgnoreGravity
-    - type: Fixtures
-      fixtures:
-        fix1:
-          shape: !type:PhysShapeCircle
-            radius: 0.35
-          density: 50
-          mask:
-            - SmallMobMask
-          layer:
-            - SmallMobLayer
-    - type: Appearance
-    - type: GenericVisualizer
-      visuals:
-        enum.DroneVisuals.Status:
-          base:
-            Off: { state: shell }
-            On: { state: drone }
-    - type: ReplacementAccent
-      accent: silicon
-    - type: Repairable
-      fuelcost: 15
-      doAfterDelay: 8
-    - type: Actions
-    - type: UnpoweredFlashlight
-      toggleAction:
-        name: action-name-toggle-light
-        description: action-description-toggle-light
-        icon: { sprite: Objects/Tools/flashlight.rsi, state: flashlight }
-        iconOn: Objects/Tools/flashlight.rsi/flashlight-on.png
-        event: !type:ToggleActionEvent
-    - type: PointLight
-      enabled: false
-      radius: 3.5
-      softness: 1
-      mask: /Textures/Effects/LightMasks/cone.png
-      autoRot: true
-    - type: Tag
-      tags:
-        - ShoesRequiredStepTriggerImmune
-        - CannotSuicide
-    - type: Hands
-      showInHands: false
-    - type: IntrinsicUI
-      uis:
-        - key: enum.LawsUiKey.Key
-          toggleAction:
-            name: action-name-show-laws
-            description: action-description-show-laws
-            icon: Structures/Wallmounts/posters.rsi/poster11_legit.png #someone wanna make new icons?
-            iconOn: Structures/Wallmounts/posters.rsi/poster11_legit.png
-            keywords: ["AI", "console", "interface", "laws", "borg"]
-            priority: -3
-            event: !type:ToggleIntrinsicUIEvent
-    - type: IntrinsicRadioReceiver
-      channels:
-        - Binary
-    - type: ActiveRadio
-      channels:
-        - Binary
